+++ src/system1/__init__.py
@@ -52,7 +52,7 @@ from system1 import (
     schema,
 )
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 
 def decide(